+++ software/tracksight/live_data/app/tasks/read_task/wireless.py
@@ -6,7 +6,7 @@ from threading import Thread
 import serial
 from generated import telem_pb2
 from logger import logger
-from broadcaster import can_msg_queue, CanMsg
+from tasks.broadcaster import can_msg_queue, CanMsg
 
 def _make_bytes(message):
 	"""
